Updated checks 50 - 124
With the exception of checks 66-69W, "then if" statements were used to avoid division by zero errors in SAS - they were just replaced with "and" here. For 66-69W, "then if" was not actually necessary, so also replaced with "and".

Checks 83W and 84 were slightly modified from the SAS file sent in April per feedback from a State agency.